crypto_levels_bhushan/backend/requirements.txt
fastapi>=0.104.0,<0.116.0
uvicorn[standard]>=0.24.0,<0.33.0
pymongo>=4.6.0,<5.0.0
motor>=3.3.0,<4.0.0
dnspython>=2.4.0,<3.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.0,<3.0.0
yfinance>=0.2.0,<1.0.0
bcrypt>=4.1.0,<5.0.0
PyJWT>=2.8.0,<3.0.0
python-multipart>=0.0.6,<1.0.0
